+++ egui/src/containers/area.rs
@@ -45,6 +45,7 @@ pub struct Area {
     pub(crate) id: Id,
     movable: bool,
     interactable: bool,
+    enabled: bool,
     order: Order,
     default_pos: Option<Pos2>,
     new_pos: Option<Pos2>,
@@ -56,6 +57,7 @@ impl Area {
             id: Id::new(id_source),
             movable: true,
             interactable: true,
+            enabled: true,
             order: Order::Middle,
             default_pos: None,
             new_pos: None,
@@ -71,6 +73,15 @@ impl Area {
         LayerId::new(self.order, self.id)
     }
 
+    /// If false, no content responds to click
+    /// and widgets will be shown grayed out.
+    /// You won't be able to move the window.
+    /// Default: `true`.
+    pub fn enabled(mut self, enabled: bool) -> Self {
+        self.enabled = enabled;
+        self
+    }
+
     /// moveable by dragging the area?
     pub fn movable(mut self, movable: bool) -> Self {
         self.movable = movable;
@@ -78,8 +89,12 @@ impl Area {
         self
     }
 
+    pub fn is_enabled(&self) -> bool {
+        self.enabled
+    }
+
     pub fn is_movable(&self) -> bool {
-        self.movable
+        self.movable && self.enabled
     }
 
     /// If false, clicks goes straight through to what is behind us.
@@ -121,6 +136,7 @@ pub(crate) struct Prepared {
     layer_id: LayerId,
     state: State,
     movable: bool,
+    enabled: bool,
 }
 
 impl Area {
@@ -130,6 +146,7 @@ impl Area {
             movable,
             order,
             interactable,
+            enabled,
             default_pos,
             new_pos,
         } = self;
@@ -149,6 +166,7 @@ impl Area {
             layer_id,
             state,
             movable,
+            enabled,
         }
     }
 
@@ -214,13 +232,15 @@ impl Prepared {
             clip_rect = clip_rect.intersect(central_area);
         }
 
-        Ui::new(
+        let mut ui = Ui::new(
             ctx.clone(),
             self.layer_id,
             self.layer_id.id,
             max_rect,
             clip_rect,
-        )
+        );
+        ui.set_enabled(self.enabled);
+        ui
     }
 
     #[allow(clippy::needless_pass_by_value)] // intentional to swallow up `content_ui`.
@@ -229,6 +249,7 @@ impl Prepared {
             layer_id,
             mut state,
             movable,
+            enabled,
         } = self;
 
         state.size = content_ui.min_rect().size();
@@ -247,6 +268,7 @@ impl Prepared {
             interact_id,
             state.rect(),
             sense,
+            enabled,
         );
 
         if move_response.dragged() && movable {